+++ src/v2/Apps/Auctions/Utils/auctionsHelpers.tsx
